[FIX]デプロイ用修正

+++ db/migrate/20200604021706_create_post_comments.rb
@@ -1,8 +1,8 @@
 class CreatePostComments < ActiveRecord::Migration[5.2]
   def change
     create_table :post_comments do |t|
-      t.references :user, foreign_key: true, type: :integer
-      t.references :team, foreign_key: true, type: :integer
+      t.references :user, foreign_key: true, type: :bigint
+      t.references :team, foreign_key: true, type: :bigint
       t.text :comment
       t.timestamps
     end
