+++ db/migrate/20200604021912_create_favorites.rb
@@ -1,8 +1,8 @@
 class CreateFavorites < ActiveRecord::Migration[5.2]
   def change
     create_table :favorites do |t|
-    	t.references :user, foreign_key: true, type: :integer
-    	t.references :team, foreign_key: true, type: :integer
+    	t.references :user, foreign_key: true, type: :bigint
+    	t.references :team, foreign_key: true, type: :bigint
     	t.timestamps
     end
   end
