+++ db/migrate/20200604022751_create_user_rooms.rb
@@ -1,8 +1,8 @@
 class CreateUserRooms < ActiveRecord::Migration[5.2]
   def change
     create_table :user_rooms do |t|
-	  t.references :user, foreign_key: true, type: :integer
-	  t.references :room, foreign_key: true, type: :integer
+	  t.references :user, foreign_key: true, type: :bigint
+	  t.references :room, foreign_key: true, type: :bigint
       t.timestamps
     end
   end
